// src/interfaces/WazeIncident.ts
export interface WazeIncident {
    city: string;
    confidence: number;
    country: string;
    county: string;
    event_class: string;
    latitude: number;
    longitude: number;
    magvar: number;
    n_thumbs_up: string;
    pub_millis: string;
    reliability: number;
    report_description: string;
    report_rating: number;
    request_millis: string;
    road_type: number;
    street: string;
    subtype: string;
    type: string;
    uuid: string;
}